+++ crates/proto/src/version/v729/info.rs
@@ -1,1 +1,1 @@
-pub const PROTOCOL_VERSION: i32 = 729;
+pub const PROTOCOL_VERSION: u16 = 729;
